Avoid running tests against Storybook v5
The previous commit updates the test harness to simulate the
installation of a '@remix/storycap' package instead of a 'storycap'
package, which requires us to introduce a Webpack alias so all of the
Storybook code points to our custom package. This is difficult to do
in Storybook v5, which structures its configuration files differently
(at least, before v5.3). Rather than spending time trying to figure out
how to make this work, we opt to skip the tests for Storybook v5, since
we're already on Storybook v6 in the Remix monorepo anyways.

See:
https://github.com/storybookjs/storybook/blob/next/MIGRATION.md#from-version-52x-to-53x.