Expose package subdirectories
In our Remix monorepo, we need to be able to access modules under
./lib/client/ and ./lib/types/, but a change upstream prevents us from
doing so because they're not listed as package entry points. In this
commit, we expose the files under each of those paths so that our
monorepo is happy once again.

See: https://nodejs.org/api/packages.html#package-entry-points